fix: add threedol style support, dev entry + card-movement routing
- Add threedol to STYLE_OPTIONS and decadeLayoutStyles (fixes player7.css not loading)
- Create src/index.js (pure JS entry, no build needed for dev)
- Add card-movement.js routing: auto-load card-movement-{style}.js via game.checkFile, fallback to card-movement-base.js
- Split card-movement.js -> card-movement-base.js + card-movement-onlineUI.js skeleton
- Add playerAddVirtualEquip to card-movement exports (merge fix)
